[add][normal]1.增加2d纹理图；2.增加Texture纹理类

+++ cpp_project_win/OpenGLWindow/src/VertexBufferLayout.h
@@ -14,8 +14,8 @@ struct VertexBufferElement
 		switch (type)
 		{
 		case GL_FLOAT:					return sizeof(GLfloat);
-		case GL_UNSIGNED_INT:		return sizeof(GLfloat);
-		case GL_UNSIGNED_BYTE:	return sizeof(GLfloat);
+		case GL_UNSIGNED_INT:		return sizeof(GLuint);
+		case GL_UNSIGNED_BYTE:	return sizeof(GLubyte);
 		}
 		ASSERT(false);
 		return 0;
